backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
pydantic-settings
python-jose
argon2-cffi
python-multipart
python-dotenv
cloudinary
geoalchemy2
asyncpg
geojson
shapely